memory-infrastructure/backend/requirements.txt
# Core
fastapi>=0.109.0
pydantic>=2.0.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Async
httpx>=0.26.0
aiohttp>=3.9.0
anyio>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.0
